simple P value
do not convert P value to -log10

+++ fastQTLmapping.h
@@ -100,7 +100,7 @@ struct linearFitRlt{
     int currentOmics2;
     float b;
     float t;
-    float p;
+    double p;
     float se; // beta / t
     float r2; 
     int nmiss;
